Move abstract const to rustc_middle::ty
